Rework HttpHandler interfaces for android
- HttpHandler is an interface which can be implemented by the client to
override the default implementation (DefaultHttpHandler)
- HttpResponse provides a generic network response fail/success methods.
        - This provides hooks to internal native calls.
- Client will wrap HttpResponse in its implementation dependent callback
        - Example DefaultHttpHandler wraps Tangram.HttpResponse in
        OkHttp3.Callback
- DefaultHttpHandler provides a default HttpHandler implementation which
clients can use if none is provided by the client app.
        - Overrides HttpHandler methods
        - Wraps HttpResponse in OkHttp3.Callback
